move notification relevant signals to correct app
