+++ src/argus/util/testing.py
@@ -1,9 +1,9 @@
 from django.db.models.signals import post_save
 
-from argus.incident.signals import send_notification
-from argus.incident.signals import background_send_notification
-from argus.incident.signals import add_planned_maintenance_tasks_covering_incident
 from argus.incident.models import Event, Incident
+from argus.notificationprofile.signals import task_send_notification
+from argus.notificationprofile.signals import task_background_send_notification
+from argus.plannedmaintenance.signals import add_planned_maintenance_tasks_covering_incident
 
 
 __all__ = [
@@ -16,12 +16,12 @@ __all__ = [
 
 
 def disconnect_signals():
-    post_save.disconnect(send_notification, Event, dispatch_uid="send_notification")
-    post_save.disconnect(background_send_notification, Event, dispatch_uid="send_notification")
+    post_save.disconnect(task_send_notification, Event, dispatch_uid="send_notification")
+    post_save.disconnect(task_background_send_notification, Event, dispatch_uid="send_notification")
     post_save.disconnect(add_planned_maintenance_tasks_covering_incident, Incident)
 
 
 def connect_signals():
-    post_save.connect(send_notification, Event, dispatch_uid="send_notification")
-    post_save.connect(background_send_notification, Event, dispatch_uid="send_notification")
+    post_save.connect(task_send_notification, Event, dispatch_uid="send_notification")
+    post_save.connect(task_background_send_notification, Event, dispatch_uid="send_notification")
     post_save.connect(add_planned_maintenance_tasks_covering_incident, Incident)
